feat(manifest): support FF-A v1.2 direct messages
Add additional messaging method parsing to indicate whether a partition
can send/receive messages via FFA_MSG_SEND_DIRECT_REQ2.

Also enforce that a manifest declaring an FF-A version < FF-A v1.2
cannot declare usage of FFA_MSG_SEND_DIRECT_REQ2.

+++ inc/hf/ffa_partition_manifest.h
@@ -158,7 +158,7 @@ struct ffa_partition_manifest {
 	struct rx_tx rxtx;
 
 	/** mandatory - direct/indirect msg or both */
-	uint8_t messaging_method;
+	uint16_t messaging_method;
 	/** mandatory - action in response to non secure interrupt */
 	uint8_t ns_interrupts_action;
 	/** optional - managed exit signaled through vIRQ */
