+++ src/vm.c
@@ -955,7 +955,7 @@ bool vm_notifications_info_get(struct vm_locked vm_locked, uint16_t *ids,
 /**
  * Checks VM's messaging method support.
  */
-bool vm_supports_messaging_method(struct vm *vm, uint8_t msg_method)
+bool vm_supports_messaging_method(struct vm *vm, uint16_t msg_method)
 {
 	return (vm->messaging_method & msg_method) != 0;
 }
